test: add macOS and Windows tests (#8)
This is to make sure we stay cross-platform.

Note that the Windows build and test is [a lot slower than Linux and Windows](https://github.com/actions/runner-images/issues/7320). Because of that, I had to change the wait in one of the tests.

Also, I'm using windows-2019, which is at least 2x faster than windows-latest: https://github.com/actions/runner-images/issues/5166

On macOS, I ran into an issue where CloseRead on a connection that was already fully read causes an error, so I needed to update the test there too. It was a huge pain to figure out what was going on.