Create PacketDialer (#14)
I'm introducing the `PacketDialer` concept that I'd like to use in the connectivity test. I added a `UDPPacketDialer`.

In the process I replaced the `PacketListenerEndpoint` with a `PacketListenerDialer`, which is more versatile. I don't think there's a big need for `PacketListenerEndpoint` yet, and you can easily implement a `PacketEndpoint` with a `PacketDialer` + remote address.